[libs] E: Move Binding for shutdown()
This commits moves the binding for the shutdown() system calls to the
syscall crate.
