--- src/libs/posix/src/sys/socket/mod.rs
@@ -1,42 +1,0 @@
-// Copyright(c) The Maintainers of Nanvix.
-// Licensed under the MIT License.
-
-//==================================================================================================
-// Imports
-//==================================================================================================
-
-use crate::errno::__errno_location;
-use ::sys::error::ErrorCode;
-use ::sysapi::ffi::c_int;
-use ::syscall::sys::{
-    socket,
-    socket::Shutdown,
-};
-
-//==================================================================================================
-// Standalone Functions
-//==================================================================================================
-
-#[unsafe(no_mangle)]
-pub extern "C" fn shutdown(sockfd: c_int, how: c_int) -> c_int {
-    ::syslog::trace!("shutdown(): sockfd={:?}, how={:?}", sockfd, how);
-
-    // Attempt to convert shutdown mode.
-    let how: Shutdown = match Shutdown::try_from(how) {
-        Ok(how) => how,
-        Err(_error) => {
-            ::syslog::error!("shutdown(): invalid shutdown mode (how={:?})", how);
-            unsafe { *__errno_location() = ErrorCode::InvalidArgument.get() };
-            return -1;
-        },
-    };
-
-    match socket::syscall::shutdown(sockfd, how) {
-        Ok(_) => 0,
-        Err(e) => {
-            ::syslog::error!("shutdown(): failed to shutdown socket {:?}", e);
-            unsafe { *__errno_location() = e.code.get() };
-            -1
-        },
-    }
-}
